backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==46.0.1
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
filetype==1.2.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.4.0
langchain==0.3.27
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.31
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.38.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
tenacity==9.1.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.25.0
